make provisioning idempotent

+++ bootstrap.sh
@@ -9,13 +9,23 @@ sudo gem install jekyll sass
 sudo npm install -g grunt
 
 # Install the Python GAE SDK
-cd /usr/local/
-sudo wget http://googleappengine.googlecode.com/files/google_appengine_1.8.9.zip -nv
-sudo unzip -q google_appengine_1.8.9.zip
-sudo rm google_appengine_1.8.9.zip
+if [ ! -d /usr/local/google_appengine ]
+then
+  cd /usr/local/
+  sudo wget http://googleappengine.googlecode.com/files/google_appengine_1.8.9.zip -nv
+  sudo unzip -q google_appengine_1.8.9.zip
+  sudo rm google_appengine_1.8.9.zip
+  echo "PATH=\$PATH:./node_modules/.bin:/usr/local/google_appengine" >> ~/.bashrc
+fi
 
-echo "PATH=\$PATH:./node_modules/.bin:/usr/local/google_appengine" >> ~/.bashrc
-
-cd /teams && npm install && cd backend && cp config_NOCOMMIT_README config_NOCOMMIT.py -u
+if [ ! -f /teams/config_NOCOMMIT.py ]
+then
+  cd /teams
+  cp config_NOCOMMIT_README config_NOCOMMIT.py
+fi
 cd /pledgeservice && npm install
-cd /authservice && cp config_NOCOMMIT_README config_NOCOMMIT.py -u
+if [ ! -f /authservice/config_NOCOMMIT.py ]
+then
+  cd /authservice
+  cp config_NOCOMMIT_README config_NOCOMMIT.py
+fi
